requirements.txt
# Finsmart ETL - AI CFO Backend
# ==============================

# Database
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# HTTP Client
requests>=2.31.0

# OpenAI SDK
openai>=1.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: FastAPI for API layer (future)
# fastapi>=0.104.0
# uvicorn>=0.24.0
